+++ example/src/App.vue
@@ -3,13 +3,5 @@ import { createDefaultApp } from "canned-shrimp";
 </script>
 
 <template>
-  <canvas :ref="el => createDefaultApp(el as HTMLCanvasElement)" class="game-canvas absolute" />
+  <canvas :ref="el => createDefaultApp(el as HTMLCanvasElement)" />
 </template>
-
-<style scoped>
-@layer page {
-  .game-canvas {
-    z-index: var(--ground);
-  }
-}
-</style>