+++ example/src/scenes/menu/Menu.scene.ts
@@ -2,14 +2,10 @@ import { defineScene, switchScene } from "canned-shrimp";
 import { Scenes } from "../../scenes.enum";
 import { defineButton } from "../../prefabs/button/button.prefab";
 
-export default function () {
-  const scene = defineScene(Scenes.Menu);
-
+export default defineScene(Scenes.Menu, (scene) => {
   const button = defineButton("Play", () => switchScene(Scenes.Game));
 
   scene.load = async () => {
     scene.addChild(button);
   };
-
-  return scene;
-}
+});