+++ example/vite.config.ts
@@ -1,9 +1,9 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
+import { cannedShrimp } from "../dist/plugin";
 
-// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [vue(), cannedShrimp()],
   envPrefix: [
     "GAME_MAP_X_LENGTH",
     "GAME_MAP_Y_LENGTH",