+++ lib/core/scene.ts
@@ -1,8 +1,21 @@
 import type { Scene } from "./types/scene";
 import { defineContainer } from "@/composables/core";
 
-export function defineScene(name: string | number): Scene {
+function createScene(name: string | number): Scene {
   const container = defineContainer(name.toString());
 
   return Object.assign(container, { sceneName: name, index: -1 });
+}
+
+export function defineScene<T>(name: Parameters<typeof createScene>["0"], sceneSetup: (scene: Scene) => T) {
+  const creator = async () => {
+    const scene = createScene(name);
+    await sceneSetup(scene);
+
+    return scene;
+  };
+
+  Reflect.set(creator, "sceneName", name);
+
+  return creator;
 }