+++ lib/core/setup.ts
@@ -1,10 +1,19 @@
 import { coreStore } from "./core-store";
 import type { SetupOption } from "./types/setup";
 
-export function setup({ keyboard: actionKey, defaultScene, pixiAppConfig, scenes }: SetupOption) {
+export function setup({ keyboard: actionKey, defaultScene, pixiAppConfig, features }: SetupOption) {
   if (actionKey) coreStore.actionKey = actionKey;
 
   coreStore.defaultScene = defaultScene;
   coreStore.pixiApplicationOptions = pixiAppConfig;
-  coreStore.scenes = scenes;
+
+  if (features) {
+    for (const key in features) {
+      if (Object.prototype.hasOwnProperty.call(features, key)) {
+        const k = key as keyof SetupOption["features"];
+        const value = features[k];
+        if (value) coreStore.features![k] = value;
+      }
+    }
+  }
 }