+++ src/scenes/Game/Game.scene.ts
@@ -2,14 +2,12 @@ import { Text } from "pixi.js";
 import { Scenes } from "../../scenes.enum";
 import { defineScene } from "@/core/scene";
 
-export default function () {
-  const scene = defineScene(Scenes.Game);
-
+export default defineScene(Scenes.Game, (scene) => {
   const title = new Text("Gme", { fill: "white" });
+
+  // const s = new Sprite(Texture.from("food"));
 
   scene.load = async () => {
     scene.addChild(title);
   };
-
-  return scene;
-}
+});