+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # canned-shrimp
 
+## 0.0.3
+
+### Patch Changes
+
+- d9f9c8e: Enabling automatic file import via Vite plugin
+- d9f9c8e: Improve the defineScene method to support types
+
 ## 0.0.2
 
 ### Patch Changes